refactor: extract task submission from AgentWorkspace
